requirements.txt
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
camelot-py==0.9.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
decorator==5.1.1
Elixir==0.7.1
et-xmlfile==1.1.0
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
numpy==1.24.3
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
pbr==5.11.1
pdfminer.six==20221105
Pillow==9.5.0
plotly==5.15.0
protobuf==4.23.3
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
PyPDF2==2.12.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz
pytz-deprecation-shim
requests==2.31.0
rich==13.4.2
six==1.16.0
smmap==5.0.0
streamlit==1.23.1
streamlit-modal
Tempita==0.5.2
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
xlrd
xlwt==0.7.2
zipp==3.15.0